Rename WrappedDefaults -> WrappedShape
